MAINT: Refactor code into separate files
