+++ mnefun/_stats.py
